+++ project/app/(tabs)/index.tsx
@@ -1,37 +1,43 @@
 import { StyleSheet } from "react-native";
-import { useEffect, useState } from "react";
-import EditScreenInfo from "../../components/EditScreenInfo";
-import { Text, View } from "../../components/Themed";
 import { Image } from "expo-image";
-import { Organic, getOrganic } from "../../db";
+import { Text, View } from "../../components/Themed";
 import { ScrollView } from "react-native-gesture-handler";
+import { useEffect, useState } from "react";
+import { Category, getCategory } from "../../db";
 
-export default function TabOneScreen() {
+export default function TabTwoScreen() {
   const [status, setStatus] = useState("");
-  const [organic, setOrganic] = useState<Organic[]>([]);
+  const [category, setCategory] = useState<Category[] | undefined>();
 
   useEffect(() => {
-    getOrganic().then((res) => {
-      setOrganic(res);
+    getCategory().then((res) => {
+      setCategory(res);
     });
-  }, [status, organic]);
+  }, [status, category]);
 
   return (
     <ScrollView>
       <View style={styles.container}>
-        {organic?.map((org) => {
+        {category?.map((categ) => {
           return (
-            <View style={styles.container} key={org.position}>
-              <Image
-                style={styles.image}
-                source={org.link}
-                contentFit="scale-down"
-                transition={1000}
-              />
-              <View>
-                <Text style={styles.title}>{org.title}</Text>
-                <Text style={styles.text}>{org.snippet}</Text>
-                <Text style={styles.text}>{org?.date}</Text>
+            <View style={styles.infoContainer} key={categ.index}>
+              <View style={{ flexGrow: 1, width: "50%" }}>
+                <Image
+                  style={styles.image}
+                  source={categ.logo}
+                  contentFit="scale-down"
+                  transition={1000}
+                />
+              </View>
+              <View
+                style={{
+                  flexGrow: 1,
+                  width: "50%",
+                  backgroundColor: "#287B51",
+                }}
+              >
+                <Text style={styles.title}>{categ.name}</Text>
+                <Text style={styles.text}>{categ.query}</Text>
               </View>
             </View>
           );
@@ -45,15 +51,7 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     alignItems: "center",
-    marginVertical: 20,
     backgroundColor: "white",
-    borderRadius: 15,
-    overflow: "hidden",
-  },
-  image: {
-    flex: 1,
-    width: "100%",
-    margin: 0,
   },
   title: {
     margin: 10,
@@ -62,5 +60,20 @@ const styles = StyleSheet.create({
     fontSize: 15,
     margin: 10,
     lineHeight: 25,
+  },
+  image: {
+    flex: 1,
+    margin: 0,
+    width: 100,
+    height: 100,
+    marginLeft: 25,
+  },
+  infoContainer: {
+    flexDirection: "row",
+    marginVertical: 20,
+    height: 160,
+    borderRadius: 15,
+    overflow: "hidden",
+    width: "80%",
   },
 });
